+++ src/Oro/Bundle/ApiBundle/Form/Extension/CustomizeFormDataExtension.php
@@ -16,8 +16,8 @@ use Symfony\Component\OptionsResolver\OptionsResolver;
 /**
  * Initializes the form builder by adding all options required to handle
  * all types of form related events dispatched in "customize_form_data" action,
- * and registers handlers for all the events, except "finish_submit".
- * The "finish_submit" event is processed by FormValidationHandler,
+ * and registers handlers for all the events, except "pre_validate" and "post_validate".
+ * The "pre_validate" and "post_validate" events are processed by FormValidationHandler,
  * because the deferred validation is used in Data API.
  * @see \Oro\Bundle\ApiBundle\Form\Extension\ValidationExtension
  * @see \Oro\Bundle\ApiBundle\Form\FormValidationHandler